requirements.txt
django
djangorestframework
django-cors-headers
pillow
django-rest-authtoken